// src/types/api/construction.ts
export type ConstructionDetail = {
  id: number;
  construction_code: string;
  construction_name: string;
  customer_id: number | string;
  east_west_division: number;
  new_repair_division: number;
  main_brand_id: number;
  sub_brand1_id: number;
  sub_brand2_id: number;
  type: number;
  clientDepartmentName: string;
  clientContactName: string;
  contractEmail: string;
  zipcode: string;
  prefecture: string;
  municipality: string;
  address_1: string;
  address_2: string;
  phone_no: string;
  structureRc: string;
  structureS: string;
  structureSrc: string;
  structureWooden: string;
  groundFloor: string;
  basementFloor: string;
  construction_content: string;
  siteArea: string;
  siteAreaInTsubo: string;
  buildingArea: string;
  buildingAreaInTsubo: string;
  totalFloorArea: string;
  totalFloorAreaInTsubo: string;
  product: string;
  schedule_construction_start_date: string;
  schedule_construction_end_date: string;
  updatedDate: string;
  dr0: string;
  dr1: string;
  dr2: string;
  estimated_amount: number;
  estimated_assumption_cost: number;
  order_amount: number;
  order_assumption_cost: number;
  work_place_score: number;
  user_review: string;
  assumption_artificial: number;
  estimated_date: string;
  order_date: string;
  contract_date: string;
  completionDate: string;
  construction_completion_date: string;
  fire_inspection_date: string;
};

/** 工事詳細 **/
export type TypeConstructionDetail = {
  id: number;
  construction_code: string;
  construction_name: string;
  construction_content: string;
  product: string;
  east_west_division: number;
  main_brand_id: string;
  sub_brand1_id: string;
  sub_brand2_id: string;
  site_area: number;
  building_area: number;
  total_floor_area: number;
  structure_rc: boolean;
  structure_s: boolean;
  structure_src: boolean;
  structure_wooden: boolean;
  ground_floor: number;
  underground_floor: number;
  zipcode: string;
  prefecture: string;
  municipality: string;
  address_1: string;
  address_2: string;
  phone_no: string;
  customer_id: string;
  customer_department_name: string;
  contractor_name: string;
  contract_email_address: string;
  order_amount: number;
  estimated_amount: number;
  consumption_tax: number;
  estimated_assumption_cost: number;
  order_assumption_cost: number;
  final_expected_cost: number;
  working_budget: number;
  assumption_artificial: number;
  cost_real_time: number;
  assessment_amount: number;
  remaining_budget: number;
  gross_profit_amount_1: number;
  gross_profit_rate_1: number;
  gross_profit_amount_2: number;
  gross_profit_rate_2: number;
  gross_profit_amount_3: number;
  gross_profit_rate_3: number;
  gross_profit_amount_4: number;
  gross_profit_rate_4: number;
  gross_profit_amount_5: number;
  gross_profit_rate_5: number;
  status: string;
  estimated_date: string;
  order_date: string;
  contract_date: string;
  schedule_construction_start_date: Date;
  schedule_construction_end_date: Date;
  construction_started_date: string;
  confirmation_application: string;
  fire_inspection_date: string;
  dr0: string;
  dr1: string;
  dr2: string;
  work_place_score: number;
  user_review: string;
  construction_completion_date: string;
  old_construction_code: string;
  new_repair_division: number;
  sales_amount: number;
  additional_increase_decrease: number;
  existing_order_amount: number;
  unordered_budget: number;
  expenses: number;
  total_expenses: number;
  assessment_turnover: number;
  assumption_gross_profit: number;
  company_name: string;
  updated_at: string;
  customer: TypeCustomer;
  main_brand: TypeBrand;
  sub_brand1: TypeBrand;
  sub_brand2: TypeBrand;
  overview: TypeConstructionOverview;
  information_setting: TypeConstructionInformationSetting;
  application_status: number;
  achievement: number;
  ordered_amount: number;
  target_gross_profit_amount: number;
  target_gross_profit_rate: number;
};